# swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API",
        "version": "1.0.0",
        "title": "Taxi"
    },
    "schemes": [
        "http"
    ],
    "host": "localhost:3000",
    "basePath": "/api",
    "paths": {
        "/login": {
            "post": {
                "summary": "authenticate",
                "description": "authenticate",
                "produces": [
                    "application/json"
                ],
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "login object",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/login"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "message": "логин успешен"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        },
        "/cars": {
            "get": {
                "summary": "Get all the cars",
                "description": "Get all the cars",
                "produces": [
                    "application/json"
                ],
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/carsResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            },
            "post": {
                "summary": "Save the new car",
                "description": "Save the new car",
                "produces": [
                    "application/json"
                ],
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "car object",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/cars"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/carsResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        },
        "/cars/{id}": {
            "get": {
                "summary": "Get car by nunber",
                "description": "Get car by number",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "number",
                        "in": "path",
                        "description": "car name that needs",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/carsResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            },
            "put": {
                "summary": "Update the car",
                "description": "Update the car",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "car id that needs to updating",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "description": "car object",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/carsResponse"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/carsResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            },
            "delete": {
                "summary": "Delete the car",
                "description": "Delete the car",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "car id that needs to be deleted",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        },
        "/drivers": {
            "get": {
                "summary": "Get all the drivers",
                "description": "Get all the drivers",
                "produces": [
                    "application/json"
                ],
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/driverResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            },
            "post": {
                "summary": "Save the new driver",
                "description": "Save the new driver",
                "produces": [
                    "application/json"
                ],
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "driver object",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/cars"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/driverResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        },
        "/drivers/{id}": {
            "get": {
                "summary": "Get driver by name",
                "description": "Get driver by name",
                "produces": [
                    "application/json"
                ],
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/driverResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            },
            "put": {
                "summary": "Update the driver",
                "description": "Update the driver",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "driver id that needs to updating",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "description": "driver object",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/driverResponse"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/driverResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            },
            "delete": {
                "summary": "Delete the driver",
                "description": "Delete the driver",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "car id that needs to be driver",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "login": {
            "type": "object",
            "properties": {
                "username": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "driver": {
            "type": "object",
            "properties": {
                "cars": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "class": {
                                "type": "string"
                            },
                            "firm": {
                                "type": "string"
                            },
                            "model": {
                                "type": "string"
                            },
                            "number": {
                                "type": "string"
                            },
                            "owner": {
                                "type": "string"
                            },
                            "year": {
                                "type": "number"
                            }
                        }
                    }
                },
                "firstName": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                },
                "age": {
                    "type": "number"
                },
                "dateBeginWork": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "reyting": {
                    "type": "number"
                },
                "city": {
                    "type": "string"
                }
            }
        },
        "driverResponse": {
            "type": "object",
            "properties": {
                "cars": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "class": {
                                "type": "string"
                            },
                            "firm": {
                                "type": "string"
                            },
                            "model": {
                                "type": "string"
                            },
                            "number": {
                                "type": "string"
                            },
                            "owner": {
                                "type": "string"
                            },
                            "year": {
                                "type": "number"
                            }
                        }
                    }
                },
                "firstName": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                },
                "age": {
                    "type": "number"
                },
                "dateBeginWork": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "reyting": {
                    "type": "number"
                },
                "city": {
                    "type": "string"
                }
            }
        },
        "cars": {
            "type": "object",
            "properties": {
                "class": {
                    "type": "string"
                },
                "firm": {
                    "type": "string"
                },
                "model": {
                    "type": "string"
                },
                "number": {
                    "type": "string"
                },
                "owner": {
                    "type": "string"
                },
                "year": {
                    "type": "number"
                }
            }
        },
        "carsResponse": {
            "type": "object",
            "properties": {
                "class": {
                    "type": "string"
                },
                "firm": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "model": {
                    "type": "string"
                },
                "number": {
                    "type": "string"
                },
                "owner": {
                    "type": "string"
                },
                "year": {
                    "type": "number"
                }
            }
        },
        "InvalidResponse": {
            "type": "object",
            "properties": {
                "statusCode": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    }
}